create a single entry point to the module
The current state of the module makes it impossible to install a munin
server with munin::host, since it depends on resources in munin::client
(because of the plugins it installs). Even if we include the
munin::client class "just before" munin::host, puppet thinks there are
dependency issues.

By moving arguments to a single point of entry we can then include all
necessary items and puppet will find the needed resources correctly.